Move types to component file
