--- frontend/src/components/Cell/cellData.ts
@@ -1,11 +1,0 @@
-type CellData =
-  | {
-      editable: true;
-      filled: boolean;
-    }
-  | {
-      editable: false;
-      content: { letter: string | null; number: number | null } | null;
-    };
-
-export default CellData;
