+++ frontend/src/components/Grid/index.tsx
@@ -1,9 +1,14 @@
 import React from 'react';
-import Cell from '../Cell';
-import CellData from '../../components/Cell/cellData';
+import Cell, { CellData } from '../Cell';
 import Corner from '../Cell/corner';
-import Mode from './mode';
 import './styles.scss';
+
+export enum Mode {
+  Answer = 'answer',
+  Draw = 'draw',
+  Erase = 'erase',
+  Puzzle = 'puzzle',
+}
 
 type Props =
   | {
