--- frontend/src/components/Grid/mode.ts
@@ -1,8 +1,0 @@
-enum Mode {
-  Answer = 'answer',
-  Draw = 'draw',
-  Erase = 'erase',
-  Puzzle = 'puzzle',
-}
-
-export default Mode;
