--- frontend/src/components/Label/labelSize.ts
@@ -1,6 +1,0 @@
-enum LabelSize {
-  Small = 'small',
-  Large = 'large',
-}
-
-export default LabelSize;
